% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pickMarble.R
\name{pickMarble}
\alias{pickMarble}
\title{Simulates the Drawing of A Marble.}
\usage{
pickMarble(iter = 10, mech = "bsc", do.bag = NULL, do.size = NULL,
  do.color = NULL)
}
\arguments{
\item{iter}{integer, the number of independent iterations (defaults to 10).}

\item{mech}{character, one  among  "bsc" (default  value), "bcs",  "sbc",
"scb", "cbs", "csb", to specify the  inner mechanism used to simulate the
drawing. The  ordering of the  first letters of  the words bag,  size and
color determines in which order the features are drawn.}

\item{do.bag}{either  NULL for no intervention on bag  (default value) or a
character  to specify  the nature  of the  intervention, either  "one" or
"two".}

\item{do.size}{either NULL for no intervention on size (default value) or a
character to  specify the nature  of the intervention, either  "large" or
"small".}

\item{do.color}{either NULL for no intervention on color (default value) or
a character  to specify the nature  of the intervention, either  "red" or
"blue".}
}
\value{
A  data.frame with iter rows  and 3 columns describing  the results of
  the iter independent samplings.
}
\description{
\code{pickMarble} simulates the drawing (with  replacement) from a box of a
marble characterized by the bag it belonged to before being poured into the
box, its size and color.  Six  inner mechanisms are available, one for each
ordering of these features.
}
\details{
\code{pickMarble} simulates the  drawing (with replacement)  of a
  marble from  a box of marbles  characterized by the bag  they belonged to
  before  being poured  into the  box (\emph{bag  one} or  \emph{bag two}),
  their  size  (\emph{large}  or  \emph{small}) and  color  (\emph{red}  or
  \emph{blue}).  Six inner mechanisms are considered, one for each ordering
  of the features  of a marble. For a generic  inner mechanism described by
  "xyz" with distinct "x", "y", "z" taken in \eqn{\{}"b", "s", "c"\eqn{\}},
  the sampling consists in \enumerate{\item drawing a bag if x=b, a size if
  x=s, a color if x=c, \item drawing  a bag if y=b, a size if
  y=s, a color if y=c, \item drawing a bag if z=b, a size if
  z=s, a color if z=c.}  Introduced
  by  J.  G.   Bennett to  discuss Simpson's  paradox, the  distribution of
  marbles is the following: \tabular{rcccccccc}{\tab bag \tab one \tab \tab
  bag \tab two \tab \tab the \tab box  \cr \tab red \tab blue \tab \tab red
  \tab blue \tab \tab  red \tab blue \cr large \tab 12 \tab  18 \tab \tab 8
  \tab 2 \tab \tab 20  \tab 20 \cr small \tab 3 \tab 7  \tab \tab 21 \tab 9
  \tab  \tab 24  \tab 16}  where \emph{the  box} contains  all the  marbles
  poured together.
}
\examples{
## draw independently 10 marbles from the box 
## with inner mechanism (1) bag, (2) size, (3) color

pickMarble(10, mech = "bsc")

## draw independently 10 marbles from bag one
## with inner mechanism (1) bag, (2) size, (3) color

pickMarble(10, mech = "bsc", do.bag = "one")

## draw independently 10 large marbles from the box
## with inner mechanism (1) bag, (2) size, (3) color

pickMarble(10, mech = "bsc", do.size = "large")

}
